Add branding to trigger button. Setup repositioning with panel.
