+++ src/front/ui/system/index.js
@@ -1,3 +1,4 @@
 export { App, AppMenu, AppMenuButton, useActiveApp } from './app'
 export { useAppFrame, AppFrame, FrameSizeButton } from './app-frame'
 export { useAppsMenu } from './apps-menu'
+export { UIToggleButton } from './ui-toggle'
